+++ src/main/java/com/onshop/shop/address/Address.java
@@ -1,0 +1,41 @@
+
+package com.onshop.shop.address;
+
+import org.hibernate.annotations.OnDelete;
+import org.hibernate.annotations.OnDeleteAction;
+
+import com.onshop.shop.user.User;
+
+import jakarta.persistence.CascadeType;
+import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.Table;
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+@Entity
+@Table(name = "address")
+public class Address {
+
+    @Id
+    @GeneratedValue
+    private Long addressId;
+
+    @ManyToOne(fetch = FetchType.LAZY) // User와 다대일 관계
+    @JoinColumn(name = "user_id", nullable = false)
+    @OnDelete(action = OnDeleteAction.CASCADE) // User 삭제 시, 관련된 Address도 삭제됨
+    private User user;
+
+    private String address1;
+    private String address2;
+    private String post;
+    private Boolean isDefault;
+}
+
