+++ src/main/java/com/onshop/shop/address/AddressDTO.java
@@ -1,12 +1,12 @@
 
-package com.onshop.shop.user;
+package com.onshop.shop.address;
 
 import lombok.Getter;
 import lombok.Setter;
 
 @Getter
 @Setter
-public class AddressDto {
+public class AddressDTO {
     private String address1;
     private String address2;
     private String post;
