+++ src/main/java/com/onshop/shop/address/AddressRepository.java
@@ -1,4 +1,4 @@
-package com.onshop.shop.user;
+package com.onshop.shop.address;
 
 import org.apache.ibatis.annotations.Param;
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -6,18 +6,20 @@ import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.transaction.annotation.Transactional;
 
-public interface AddressRepository extends JpaRepository<Address, Integer> {
+import com.onshop.shop.user.User;
+
+public interface AddressRepository extends JpaRepository<Address, Long> {
 
     @Transactional
     void deleteByUser(User user);
 
     @Transactional
-    void deleteByUserUserId(Integer userId);
+    void deleteByUserUserId(Long userId);
 
     long countByUser(User user);
     
     @Modifying
     @Transactional
     @Query("DELETE FROM Address a WHERE a.user.userId = :userId")
-    void deleteAllByUserId(@Param("userId") Integer userId);
+    void deleteAllByUserId(@Param("userId") Long userId);
 }
