+++ src/main/java/com/onshop/shop/article/Article.java
@@ -7,16 +7,19 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 import java.time.LocalDateTime;
 
+import org.hibernate.annotations.OnDelete;
+import org.hibernate.annotations.OnDeleteAction;
+
 @Entity
 @Getter
 @Setter
 @NoArgsConstructor
 @Table(name = "article")
-public class ArticleEntity {
+public class Article {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int articleId;
+    private Long articleId;
 
     @Column(nullable = false, length = 255)
     private String articleName;
@@ -27,7 +30,8 @@ public class ArticleEntity {
     // UserEntity와 연결, user_id가 NULL이 될 수도 있도록 설정
     @ManyToOne
     @JoinColumn(name = "user_id", foreignKey = @ForeignKey(name = "fk_article_user"), nullable = true)
-    private User userEntity;
+    @OnDelete(action = OnDeleteAction.SET_NULL) // User 삭제 시 user_id를 NULL로 설정
+    private User user;
 
     @Column(columnDefinition = "TIMESTAMP DEFAULT CURRENT_TIMESTAMP")
     private LocalDateTime writtenDate;
