+++ src/main/java/com/onshop/shop/article/ArticleRepository.java
@@ -2,5 +2,5 @@ package com.onshop.shop.article;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ArticleRepository extends JpaRepository<ArticleEntity, Integer> {
+public interface ArticleRepository extends JpaRepository<Article, Long> {
 }
