+++ src/main/java/com/onshop/shop/article/ArticleServiceImpl.java
@@ -23,7 +23,7 @@ public class ArticleServiceImpl implements ArticleService {
 
     // ✅ 특정 게시물 조회
     @Override
-    public ArticleDTO getArticleById(int id) {
+    public ArticleDTO getArticleById(Long id) {
         return articleRepository.findById(id)
                 .map(ArticleDTO::fromEntity) // Entity → DTO 변환
                 .orElseThrow(() -> new RuntimeException("게시물을 찾을 수 없습니다."));
@@ -32,25 +32,25 @@ public class ArticleServiceImpl implements ArticleService {
     // ✅ 게시물 생성
     @Override
     public ArticleDTO createArticle(ArticleDTO articleDTO) {
-        ArticleEntity articleEntity = new ArticleEntity();
+        Article articleEntity = new Article();
         articleEntity.setArticleName(articleDTO.getArticleName());
         articleEntity.setArticle(articleDTO.getArticle());
         // UserEntity 설정 (userId가 있을 경우만)
         if (articleDTO.getUserId() > 0) {
             User user = new User();
             user.setUserId(articleDTO.getUserId());
-            articleEntity.setUserEntity(user);
+            articleEntity.setUser(user);
         }
         articleEntity.setWrittenDate(articleDTO.getWrittenDate());
 
-        ArticleEntity savedArticle = articleRepository.save(articleEntity);
+        Article savedArticle = articleRepository.save(articleEntity);
         return ArticleDTO.fromEntity(savedArticle); // 저장 후 DTO 반환
     }
 
     // ✅ 게시물 수정 (부분 업데이트 지원)
     @Override
-    public ArticleDTO updateArticle(int id, ArticleDTO articleDTO) {
-        ArticleEntity articleEntity = articleRepository.findById(id)
+    public ArticleDTO updateArticle(Long id, ArticleDTO articleDTO) {
+        Article articleEntity = articleRepository.findById(id)
                 .orElseThrow(() -> new RuntimeException("게시물을 찾을 수 없습니다."));
 
         // 변경할 필드만 업데이트
@@ -61,13 +61,13 @@ public class ArticleServiceImpl implements ArticleService {
             articleEntity.setArticle(articleDTO.getArticle());
         }
 
-        ArticleEntity updatedArticle = articleRepository.save(articleEntity);
+        Article updatedArticle = articleRepository.save(articleEntity);
         return ArticleDTO.fromEntity(updatedArticle); // 업데이트 후 DTO 반환
     }
 
     // ✅ 게시물 삭제
     @Override
-    public void deleteArticle(int id) {
+    public void deleteArticle(Long id) {
         if (!articleRepository.existsById(id)) {
             throw new RuntimeException("삭제할 게시물을 찾을 수 없습니다.");
         }
