+++ src/main/java/com/onshop/shop/cart/CartDTO.java
@@ -1,5 +1,6 @@
 package com.onshop.shop.cart;
 
+
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
@@ -7,9 +8,9 @@ import lombok.Data;
 @AllArgsConstructor
 public class CartDTO {
     private Long cartId;
-    private Long productId;
     private int quantity;
-    private int userId;
+	private Long userId;
+    private Long productId;
     private String productName; // ✅ product에서 상품명 추가
     private int productPrice;   // ✅ product에서 상품 가격 추가
 }
