+++ src/main/java/com/onshop/shop/reply/ReplyController.java
@@ -16,14 +16,14 @@ public class ReplyController {
 
     // ✅ 특정 게시물의 댓글 조회
     @GetMapping
-    public ResponseEntity<List<ReplyDTO>> getRepliesByArticleId(@PathVariable int articleId) {
+    public ResponseEntity<List<ReplyDTO>> getRepliesByArticleId(@PathVariable Long articleId) {
         List<ReplyDTO> replies = replyService.getRepliesByArticleId(articleId);
         return ResponseEntity.ok(replies);
     }
 
     // ✅ 댓글 작성
     @PostMapping
-    public ResponseEntity<ReplyDTO> createReply(@PathVariable int articleId, @RequestBody ReplyDTO replyDTO) {
+    public ResponseEntity<ReplyDTO> createReply(@PathVariable Long articleId, @RequestBody ReplyDTO replyDTO) {
         replyDTO.setArticleId(articleId); // 게시물 ID 설정
         ReplyDTO createdReply = replyService.createReply(replyDTO);
         return ResponseEntity.ok(createdReply);
@@ -31,7 +31,7 @@ public class ReplyController {
 
     // ✅ 댓글 삭제
     @DeleteMapping("/{commentId}")
-    public ResponseEntity<Void> deleteReply(@PathVariable int articleId, @PathVariable int commentId) {
+    public ResponseEntity<Void> deleteReply(@PathVariable Long articleId, @PathVariable Long commentId) {
         replyService.deleteReply(commentId);
         return ResponseEntity.noContent().build();
     }
