+++ src/main/java/com/onshop/shop/security/SecurityConfig.java
@@ -30,7 +30,7 @@ public class SecurityConfig {
             .sessionManagement(session -> session.sessionCreationPolicy(SessionCreationPolicy.STATELESS))
             .authorizeHttpRequests(auth -> auth
                 .requestMatchers("/api/**","/auth/**", "/category",
-                	      "/products/**",    "/products", "/category/**", "/products/**", "/cart/**", "/favicon.ico","/uploads/**"
+                	      "/products/**",    "/products", "/category/**", "/products/**","/cart", "/cart/**", "/favicon.ico","/uploads/**"
                         ,"/store/**","/seller/**","/seller","/seller/info/**"
                 		)
