+++ src/main/java/com/onshop/shop/seller/products/SellerProductsService.java
@@ -8,4 +8,8 @@ public interface SellerProductsService {
     void removeProducts(SellerProductIdsDTO productsIds); 	// 상품 삭제
     void updateProducts(Long productId, SellerProductsRequestDTO product); // 상품 정보 수정
     List<SellerProductsDTO> searchProducts(String search, int page, int size);
+    
+    
+    
+    
 }
