+++ src/main/java/com/onshop/shop/store/Seller.java
@@ -39,6 +39,9 @@ public class Seller {
     @Column(name = "navigation_id", nullable = false)
     private Integer navigationId;  // 네비게이션 ID (페이지 이동 방식)
 
+    @Column(name = "seller_menubar_color", length = 7, nullable = true)
+    private String sellerMenubarColor;  // ✅ 메뉴바 색상 (예: "#808080")
+
     @Column(name = "created_at", columnDefinition = "DATETIME DEFAULT CURRENT_TIMESTAMP")
     private LocalDateTime createdAt;  // 생성일
 
@@ -47,15 +50,25 @@ public class Seller {
 
     public Seller() {}
 
-    public Seller(Long sellerId, Long userId, String storename, String description, Integer headerId, Integer menuBarId, Integer navigationId, LocalDateTime createdAt, LocalDateTime updatedAt) {
+    public Seller(Long sellerId, Long userId, String storename, String description, Integer headerId, Integer menuBarId, Integer navigationId, String sellerMenubarColor, LocalDateTime createdAt, LocalDateTime updatedAt) {
         this.sellerId = sellerId;
-    	this.userId = userId;
+        this.userId = userId;
         this.storename = storename;
         this.description = description;
         this.headerId = headerId;
         this.menuBarId = menuBarId;
         this.navigationId = navigationId;
+        this.sellerMenubarColor = sellerMenubarColor; // ✅ 추가
         this.createdAt = createdAt;
         this.updatedAt = updatedAt;
+    }
+
+    // ✅ Getter 및 Setter 추가
+    public String getSellerMenubarColor() {
+        return sellerMenubarColor;
+    }
+
+    public void setSellerMenubarColor(String sellerMenubarColor) {
+        this.sellerMenubarColor = sellerMenubarColor;
     }
 }
