+++ src/main/java/com/onshop/shop/store/StoreProductController.java
@@ -9,7 +9,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.onshop.shop.products.Product;
+import com.onshop.shop.product.Product;
 
 @RestController
 @CrossOrigin(origins = "http://localhost:5173")
