+++ src/main/java/com/onshop/shop/store/StoreProductDTO.java
@@ -11,6 +11,4 @@ public class StoreProductDTO {
     private String description;
     private int price;
     private String imageUrl;
-    private Long categoryId;  // FK 매핑
-    private Long storeId;  // ✅ Store 엔티티 대신 storeId만 포함
 }
