+++ src/main/java/com/onshop/shop/store/StoreProductRepository.java
@@ -1,8 +1,10 @@
 
 package com.onshop.shop.store;
 
-import com.onshop.shop.products.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
+
+import com.onshop.shop.product.Product;
+
 import java.util.List;
 
 public interface StoreProductRepository extends JpaRepository<Product, Long> {
