--- src/main/java/com/onshop/shop/user/Address.java
@@ -1,27 +1,0 @@
-
-package com.onshop.shop.user;
-
-import jakarta.persistence.*;
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-@Entity
-@Table(name = "addresses")
-public class Address {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer addressId;
-
-    @ManyToOne
-    @JoinColumn(name = "user_id", nullable = false)
-    private User user;
-
-    private String address1;
-    private String address2;
-    private String post;
-    private Boolean isDefault;
-}
-
