+++ src/main/java/com/onshop/shop/user/UserDTO.java
@@ -3,6 +3,8 @@ package com.onshop.shop.user;
 
 import java.util.List;
 
+import com.onshop.shop.address.AddressDTO;
+
 import lombok.Getter;
 import lombok.Setter;
 
@@ -18,6 +20,6 @@ public class UserDTO {
     private String role;
     
     // 주소 정보 리스트
-    private List<AddressDto> addresses;
+    private List<AddressDTO> addresses;
 }
