--- src/main/java/com/onshop/shop/user/UserResitory.java
@@ -1,6 +1,0 @@
-package com.onshop.shop.user;
-
-public class UserResitory {
-	
-
-}
